worker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.7.4
paddlepaddle==2.6.1
paddleocr==2.8.1
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
google-generativeai==0.7.2
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
